pyproject.toml
[tool.poetry]
name = "resolucion"
version = "0.1.0"
description = ""
authors = ["Mariano Re <marianoreinfo@gmail.com>"]
packages = [{include = "project"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.2"
python-decouple = "^3.6"
apache-airflow = "^2.3.2"

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"
pylint = "^2.6.0"
pre-commit = "^2.7.1"
black = "^22.3.0"
flake8 = "^3.8.4"
isort = "^5.6.4"
pre-commit-hooks = "^3.3.0"
pytest = "^6.1.1"
shellcheck-py = "^0.7.1.1"
pandas-vet = "^0.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
